Refactored LoginSelectorList into dumb component
